- Jorrit fixed a bug in csObjectModel. It would use Put() instead 	  of PutUnique() for the hash of triangle data meshes. The 	  consequence of that was that it was not possible to replace 	  a triangle mesh once it was set. A further consequence of that 	  was that it genmeshes with submeshes didn't have working 	  collision detection.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@26770 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/genmesh/object/submeshes.cpp
@@ -243,7 +243,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(Genmesh)
   {
     if (!factory) return 0;
     CacheTriangles ();
-    return (int)triangleCache.GetSize ();
+    return triangleCache.GetSize ();
   }
   csTriangle* SubMeshesTriMesh::GetTriangles ()
   {
